backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "langchain>=1.2.18",
    "langchain-anthropic>=1.4.3",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.2.1",
    "langfuse>=4.6.1",
    "langgraph>=1.1.10",
    "python-dotenv>=1.2.2",
    "uvicorn[standard]>=0.32.0",
]
